requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
APScheduler==3.9.1.post1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.9.3
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
black==22.1.0
bleach==4.1.0
blinker==1.4
boto==2.49.0
boto3==1.26.32
botocore==1.29.32
Brotli==1.0.9
CacheControl==0.12.11
cached-property==1.5.1
cachetools==5.2.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.4.0
cliapp==1.20180812.1
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cloud-init==22.3.4
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
cssutils==2.6.0
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
discord.py==2.1.0
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
executing==0.8.2
filelock==3.8.0
Fiona==1.8.22
firebase-admin==6.0.1
flake8==4.0.1
Flask==2.2.2
Flask-Compress==1.13
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.3
future==0.18.2
fuzzywuzzy==0.18.0
geocoder==1.38.1
geographiclib==2.0
geopandas==0.12.1
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
googlesearch-python==1.1.0
googletrans==3.0.0
greenlet==2.0.1
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.9.0
h2==3.2.0
herokuify-scrapyd==1.0
hiredis==2.1.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httplib2==0.21.0
httpx==0.13.3
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
imbalanced-learn==0.8.1
importlib-metadata==5.1.0
incremental==16.10.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
Levenshtein==0.20.8
libretranslatepy==2.1.1
llvmlite==0.38.0
logparser==0.8.2
lxml==4.8.0
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==4.2.0
msgpack==0.6.2
multidict==6.0.3
munch==2.5.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
nltk==3.8.1
notebook==6.4.8
numba==0.55.0
numpy==1.19.3
oauthlib==3.2.2
outcome==1.1.0
packaging==22.0
pandas==1.3.3
pandocfilters==1.5.0
parsel==1.7.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pbr==5.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pip-autoremove==0.10.0
pipenv==2022.11.30
platformdirs==2.4.1
playwright==1.28.0
pluggy==1.0.0
premailer==3.10.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
Protego==0.2.1
proto-plus==1.22.1
protobuf==3.20.3
proxyscrape==0.3.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pycountry==22.3.5
pycparser==2.21
pydeck==0.8.0
PyDispatcher==2.0.6
pydocstyle==6.1.1
pyee==9.0.4
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.6.0
pylint==2.12.2
pymacaroons==0.13.0
pymongo==4.3.3
Pympler==1.0.1
PyNaCl==1.3.0
pynvim==0.4.1
pyOpenSSL==22.1.0
pyparsing==2.4.6
pyproj==3.4.0
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
pytesseract==0.3.10
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
python-Levenshtein==0.20.8
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
python-scrapyd-api==2.1.2
python-slugify==7.0.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
queuelib==1.6.2
rapidfuzz==2.13.6
ratelim==0.1.6
redis==4.4.0
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
retrying==1.3.4
rfc3986==1.5.0
rfc3987==1.3.8
rich==13.0.1
rope==0.22.0
rsa==4.9
s3transfer==0.6.0
schematics==2.1.1
scikit-learn==1.0
scipy==1.7.1
scrapeops-scrapy-proxy-sdk==1.0
scraperapi-sdk==0.2.2
scrapinghub==2.4.0
Scrapy==2.7.1
Scrapy-Cookies==0.3
scrapy-playwright==0.0.24
scrapy-proxies==0.4
scrapy-proxy-pool==0.1.9
scrapy-rotating-proxies==0.6.2
scrapy-user-agents==0.1.1
scrapyd==1.3.0
scrapyd-client @ git+https://github.com/scrapy/scrapyd-client.git@ceae109099c88a7446a4b40bb715c39d2b7148e0
scrapydweb @ git+https://github.com/my8100/scrapydweb.git@7a3b81dba2cba4279c9465064a693bb277ac20e9
seaborn==0.11.2
SecretStorage==2.3.1
selenium==4.1.3
semver==2.13.0
Send2Trash==1.8.0
sentry-sdk==1.12.0
service-identity==18.1.0
Shapely==1.8.5.post1
simplejson==3.16.0
six==1.16.0
sktime==0.9.0
slackclient==2.9.4
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sos==4.4
soupsieve==2.3.2
spidermon==1.17.0
SQLAlchemy==1.4.45
ssh-import-id==5.10
stack-data==0.1.4
statsmodels==0.12.1
streamlit==1.16.0
strict-rfc3339==0.7
strsimpy==0.2.1
systemd-python==234
terminado==0.13.1
testpath==0.5.0
testresources==2.0.1
text-unidecode==1.3
textblob==0.17.1
texttable==1.6.2
thefuzz==0.19.0
threadpoolctl==3.0.0
tldextract==3.4.0
toml==0.10.2
tomli==2.0.0
toolz==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
translate==3.6.1
trio==0.20.0
trio-websocket==0.9.2
ttystatus==0.38
tweepy==4.12.1
Twisted==18.9.0
typing==3.7.4.3
typing_extensions==4.0.1
tzdata==2022.7
tzlocal==4.2
ua-parser==0.16.1
uberegg==0.1.1
ubuntu-advantage-tools==27.11.3
ufw==0.36
ujson==5.1.0
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.13
user-agent==0.1.10
user-agents==2.2.0
utils==1.0.1
validators==0.20.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
w3lib==2.1.0
wadllib==1.3.3
watchdog==2.2.1
wcwidth==0.2.5
webcolors==1.12
webdriver-manager==3.5.4
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==2.2.2
widgetsnbextension==3.5.2
wrapt==1.13.3
wsproto==1.1.0
xgboost==1.7.2
yapf==0.32.0
yarl==1.8.2
zipp==1.0.0
zope.interface==5.5.2
